backend/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1

# Existing dependencies
autopep8==2.0.4
bandit==1.7.5

# ML/DL frameworks
scikit-learn==1.3.2
xgboost==2.0.3
torch==2.1.2
transformers==4.36.2

# Embeddings
openai==1.6.1
sentence-transformers==2.2.2

# Authentication
PyJWT>=2.8.0
email-validator==2.1.0
bcrypt==4.1.2
requests-oauthlib==1.3.1


# Vector DB
faiss-cpu>=1.8.0

# AST/Static Analysis
tree-sitter==0.20.4
semgrep==1.52.0
pylint==3.0.3
ruff==0.1.9
radon==6.0.1

# LLM (optional - choose one)
anthropic==0.8.1
mistralai==1.0.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.2
pandas==1.5.3

# Testing
pytest==7.4.3
pytest-mock==3.12.0

# Reporting
markdown==3.5.1
reportlab==4.0.7

# Web
Jinja2==3.1.2

# Vector DB (for Q&A system)
qdrant-client==1.7.0

# Code Parsing (multi-language support)
tree-sitter==0.20.4
# py-tree-sitter-languages - Not available on PyPI, optional for Java/JS parsing

# Security Analysis
detect-secrets==1.4.0
osv>=0.0.1

# Scheduling (for monitoring)
APScheduler==3.10.4

# Enhanced embeddings (optional - for Codestral)
# Mistral API already included above for Codestral Embed support
